moving adapters, classes, and services into their own folders
